Added checks for UseDatabaseQuotaDefaults
Mailboxes can be set to use Mailbox Quotas or Database Quotas. I made
some changes so the report shows if the Database Quotas are being used
(UsesDBDefaults).  If true the report will calculate the percentage used
of the mailbox based on database quota size, and if false use the
mailbox quota size.  This will help in situations where there is a mix
of database quotas and mailbox quotas being used for users.